Make downloads conditionals

+++ drush_ovh_install.sh
@@ -1,15 +1,19 @@
 #!/bin/bash
 
-# Download drush
-wget https://github.com/drush-ops/drush/archive/6.4.0.tar.gz
-# Unpack it
-tar xvzf 6.4.0.tar.gz
-# Rename it
-mv drush-6.4.0 drush
-# Clean
-rm 6.4.0.tar.gz
+# Download drush if needed
+if [ ! -d drush ]
+then
+  # Download
+  wget https://github.com/drush-ops/drush/archive/6.4.0.tar.gz
+  # Unpack
+  tar xvzf 6.4.0.tar.gz
+  # Rename
+  mv drush-6.4.0 drush
+  # Clean
+  rm 6.4.0.tar.gz
+fi
 
-# Download complementary files
+# Download complementary files if needed
 if [ ! -f .drush_bashrc ]
 then
   wget https://raw.githubusercontent.com/DuaelFr/drush-ovh/master/.drush_bashrc
@@ -19,11 +23,9 @@ then
   wget https://raw.githubusercontent.com/DuaelFr/drush-ovh/master/drush_prepend.php
 fi
 
-# Add .drush_bashrc call in .bashrc
-if grep -Fxq ".drush_bashrc" .bashrc
+# Add .drush_bashrc call in .bashrc if needed
+if ! grep -Fxq ".drush_bashrc" .bashrc
 then
-  echo ".drush_bashrc file already referenced in .bashrc"
-else
   echo "" >> .bashrc
   echo "source .drush_bashrc" >> .bashrc
 fi
